code clean-up
moved the code for ViewSmsTestSuite to separate file

left some comments though, so you can check what's going on far easier than having to decipher each one of the tests
